+++ server/api/chefs/delete/chef.delete.js
@@ -1,31 +1,41 @@
+import prisma from "~/utils/prisma.js";
+import {getServerSession} from "#auth";
 
-export default defineEventHandler(async function(event){
-  const body = await readBody(event);
+export default defineEventHandler(async function (event) {
+    const body = await readBody(event);
+    const session = await getServerSession(event);
 
+    const existingChef = await prisma.chef.findUnique({
+        where: {
+            id: body.chefID
+        }
+    });
 
-  const existingChef = await prisma.chef.findUnique({
-    where: {
-      id: body.chefID
+    if (!existingChef) {
+        return {
+            success: false,
+            error: "Chef doesn't exist."
+        }
     }
-  });
 
-  if (!existingChef) {
+    if (session.user.role !== "ADMIN") {
+        return {
+            success: false,
+            message: "You can't delete chef, because you are not ADMIN."
+        }
+    }
+
+
+    const deletedChef = await prisma.chef.delete({
+        where: {
+            id: body.chefID
+        },
+    })
+
     return {
-      success: false,
-      error: "Chef doesn't exist."
+        success: true,
+        statusCode: 200,
+        deletedChef: deletedChef
     }
-  }
-
-  const deletedChef = await prisma.chef.delete({
-    where: {
-      id: body.chefID
-    },
-  })
-
-  return {
-    success: true,
-    statusCode: 200,
-    deletedChef: deletedChef
-  }
 
 })
